example-robot/requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
flexcache==0.3
flexparser==0.4
iniconfig==2.3.0
invoke==2.2.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
paramiko==4.0.0
phoenix6==25.4.2
photonlibpy==2026.0.1
Pint==0.25.2
platformdirs==4.5.0
pluggy==1.6.0
pycparser==2.23
pyfrc==2025.1.0
Pygments==2.19.2
PyNaCl==1.6.1
pynetconsole==2.0.4
pyntcore==2025.3.2.2
pytest==9.0.1
pytest-reraise==2.1.2
robotpy==2025.3.2.2
robotpy-apriltag==2025.3.2.2
robotpy-cli==2024.0.0
robotpy-commands-v2==2025.3.2
robotpy-cscore==2025.3.2.2
robotpy-ctre==2025.0.3.1
robotpy-hal==2025.3.2.2
robotpy-halsim-ds-socket==2025.3.2.2
robotpy-halsim-gui==2025.3.2.2
robotpy-halsim-ws==2025.3.2.2
robotpy-installer==2025.1.1
robotpy-native-apriltag==2025.3.2
robotpy-native-ntcore==2025.3.2
robotpy-native-wpihal==2025.3.2
robotpy-native-wpilib==2025.3.2
robotpy-native-wpimath==2025.3.2
robotpy-native-wpinet==2025.3.2
robotpy-native-wpiutil==2025.3.2
robotpy-navx==2025.0.1.2
robotpy-pathplannerlib==2025.2.7
robotpy-rev==2025.0.3.1
robotpy-romi==2025.3.2.2
robotpy-wpilib-utilities==2025.1.0
robotpy-wpimath==2025.3.2.2
robotpy-wpinet==2025.3.2.2
robotpy-wpiutil==2025.3.2.2
robotpy-xrp==2025.3.2.2
setuptools==80.9.0
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
wpilib==2025.3.2.2
